apps/api/pyproject.toml
[project]
name = "agentverse-api"
version = "0.1.0"
description = "AgentVerse API - AI Agent Simulation Platform"
authors = [
    { name = "AgentVerse Team", email = "team@agentverse.ai" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "redis>=5.0.1",
    "celery[redis]>=5.3.6",
    "openai>=1.10.0",
    "numpy>=1.26.3",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "sentry-sdk[fastapi]>=1.39.0",
    # Object Storage (project.md §5.4)
    "boto3>=1.34.0",
    # Observability (project.md §11 Phase 9)
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.43b0",
    "opentelemetry-instrumentation-redis>=0.43b0",
    "opentelemetry-instrumentation-celery>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-exporter-prometheus>=0.43b0",
    "psutil>=5.9.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    # Load Testing (project.md §10.2)
    "locust>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
